fix(network) split form into components
