+++ src/pages/networks/forms/NetworkForm.tsx
@@ -1,13 +1,9 @@
-import React, { FC, ReactNode, useEffect, useState } from "react";
+import React, { FC, useEffect, useState } from "react";
 import {
-  CheckboxInput,
   Col,
   Form,
-  Input,
   Notification,
   Row,
-  Select,
-  Textarea,
   useNotify,
 } from "@canonical/react-components";
 import {
@@ -29,14 +25,12 @@ import NetworkFormMenu, {
 import { FormikProps } from "formik/dist/types";
 import { updateMaxHeight } from "util/updateMaxHeight";
 import useEventListener from "@use-it/event-listener";
-import { useSettings } from "context/useSettings";
-import Loader from "components/Loader";
 import YamlForm from "pages/instances/forms/YamlForm";
-import IpAddressSelector from "pages/networks/forms/IpAddressSelector";
-import ConfigurationTable from "pages/networks/forms/ConfigurationTable";
-import { getConfigurationRow } from "pages/networks/forms/ConfigurationRow";
-import { supportsFanNetwork, supportsOvnNetwork } from "util/settings";
-import NetworkSelector from "pages/networks/forms/NetworkSelector";
+import NetworkFormMain from "pages/networks/forms/NetworkFormMain";
+import NetworkFormBridge from "pages/networks/forms/NetworkFormBridge";
+import NetworkFormDns from "pages/networks/forms/NetworkFormDns";
+import NetworkFormIpv4 from "pages/networks/forms/NetworkFormIpv4";
+import NetworkFormIpv6 from "pages/networks/forms/NetworkFormIpv6";
 
 export interface NetworkFormValues {
   readOnly: boolean;
@@ -121,31 +115,12 @@ interface Props {
 const NetworkForm: FC<Props> = ({ formik, getYaml, project }) => {
   const notify = useNotify();
   const [section, setSection] = useState(MAIN_CONFIGURATION);
-  const { data: settings, isLoading: isSettingsLoading } = useSettings();
-  const hasOvn = supportsOvnNetwork(settings);
-  const hasFan = supportsFanNetwork(settings);
 
   const updateFormHeight = () => {
     updateMaxHeight("form-contents", "p-bottom-controls");
   };
   useEffect(updateFormHeight, [notify.notification?.message, section]);
   useEventListener("resize", updateFormHeight);
-
-  if (isSettingsLoading) {
-    return <Loader />;
-  }
-
-  const getFormProps = (id: "type" | "network" | "name" | "description") => {
-    return {
-      id: id,
-      name: id,
-      onBlur: formik.handleBlur,
-      onChange: formik.handleChange,
-      value: formik.values[id] ?? "",
-      error: formik.touched[id] ? (formik.errors[id] as ReactNode) : null,
-      placeholder: `Enter ${id.replaceAll("_", " ")}`,
-    };
-  };
 
   return (
     <Form
@@ -160,515 +135,12 @@ const NetworkForm: FC<Props> = ({ formik, getYaml, project }) => {
       <Row className="form-contents" key={section}>
         <Col size={12}>
           {section === MAIN_CONFIGURATION && (
-            <React.Fragment>
-              <Row>
-                <Col size={8}>
-                  <Select
-                    {...getFormProps("type")}
-                    label="Type"
-                    help={
-                      <>
-                        Bridge (FAN) is only available on ubuntu, OVN needs to
-                        be configured in LXD as{" "}
-                        <code>network.ovn.northbound_connection</code>{" "}
-                        <a
-                          href="https://documentation.ubuntu.com/lxd/en/latest/howto/network_ovn_setup/#set-up-a-lxd-cluster-on-ovn"
-                          target="_blank"
-                          rel="noreferrer"
-                        >
-                          Learn how to set up a LXD cluster on OVN
-                        </a>
-                      </>
-                    }
-                    required
-                    options={[
-                      {
-                        label: "Bridge (standard)",
-                        value: "bridge-standard",
-                      },
-                      {
-                        label: "Bridge (FAN)",
-                        value: "bridge-fan",
-                        disabled: !hasFan,
-                      },
-                      {
-                        label: "OVN",
-                        value: "ovn",
-                        disabled: !hasOvn,
-                      },
-                    ]}
-                    onChange={(e) => {
-                      if (e.target.value === "bridge-standard") {
-                        formik.setFieldValue("type", "bridge");
-                        formik.setFieldValue("bridge_mode", "standard");
-                        formik.setFieldValue("fan_type", undefined);
-                        formik.setFieldValue("fan_overlay_subnet", undefined);
-                        formik.setFieldValue("fan_underlay_subnet", undefined);
-                        formik.setFieldValue("network", undefined);
-                        formik.setFieldValue("ipv4_l3only", undefined);
-                        formik.setFieldValue("ipv6_l3only", undefined);
-                      }
-                      if (e.target.value === "bridge-fan") {
-                        formik.setFieldValue("type", "bridge");
-                        formik.setFieldValue("bridge_mode", "fan");
-                        formik.setFieldValue("ipv4_address", undefined);
-                        formik.setFieldValue("ipv6_address", undefined);
-                        formik.setFieldValue("ipv6_nat", undefined);
-                        formik.setFieldValue("network", undefined);
-                        formik.setFieldValue("ipv4_l3only", undefined);
-                        formik.setFieldValue("ipv6_l3only", undefined);
-                      }
-                      if (e.target.value === "ovn") {
-                        formik.setFieldValue("type", "ovn");
-                        formik.setFieldValue("bridge_mode", undefined);
-                        formik.setFieldValue("fan_type", undefined);
-                        formik.setFieldValue("fan_overlay_subnet", undefined);
-                        formik.setFieldValue("fan_underlay_subnet", undefined);
-                        formik.setFieldValue("bridge_driver", undefined);
-                        formik.setFieldValue("dns_mode", undefined);
-                        formik.setFieldValue("ipv4_dhcp_expiry", undefined);
-                        formik.setFieldValue("ipv4_dhcp_ranges", undefined);
-                        formik.setFieldValue("ipv4_ovn_ranges", undefined);
-                        formik.setFieldValue("ipv6_dhcp_expiry", undefined);
-                        formik.setFieldValue("ipv6_dhcp_ranges", undefined);
-                        formik.setFieldValue("ipv6_ovn_ranges", undefined);
-                      }
-                    }}
-                    value={
-                      formik.values.type === "bridge"
-                        ? `bridge-${formik.values.bridge_mode ?? "standard"}`
-                        : formik.values.type
-                    }
-                    disabled={formik.values.readOnly}
-                  />
-                  <Input
-                    {...getFormProps("name")}
-                    type="text"
-                    label="Name"
-                    required
-                    disabled={formik.values.readOnly}
-                  />
-                  <Input
-                    {...getFormProps("description")}
-                    type="text"
-                    label="Description"
-                    disabled={formik.values.readOnly}
-                  />
-                  {formik.values.type === "ovn" && (
-                    <NetworkSelector
-                      props={getFormProps("network")}
-                      project={project}
-                      isDisabled={formik.values.readOnly}
-                    />
-                  )}
-                </Col>
-              </Row>
-              <ConfigurationTable
-                rows={[
-                  ...(formik.values.bridge_mode !== "fan"
-                    ? [
-                        getConfigurationRow({
-                          formik: formik,
-                          name: "ipv4_address",
-                          label: "IPv4 Address",
-                          defaultValue: "auto",
-                          children: (
-                            <IpAddressSelector
-                              id="ipv4_address"
-                              address={formik.values.ipv4_address}
-                              setAddress={(value) => {
-                                formik.setFieldValue("ipv4_address", value);
-
-                                if (value === "none") {
-                                  const nullFields = [
-                                    "ipv4_nat",
-                                    "ipv4_dhcp",
-                                    "ipv4_dhcp_expiry",
-                                    "ipv4_dhcp_ranges",
-                                  ];
-                                  nullFields.forEach((field) =>
-                                    formik.setFieldValue(field, undefined)
-                                  );
-                                }
-                              }}
-                            />
-                          ),
-                        }),
-                      ]
-                    : []),
-
-                  ...(formik.values.bridge_mode !== "fan" &&
-                  formik.values.ipv4_address !== "none"
-                    ? [
-                        getConfigurationRow({
-                          formik: formik,
-                          name: "ipv4_nat",
-                          label: "Ipv4 NAT",
-                          defaultValue: "",
-                          children: (
-                            <CheckboxInput
-                              label="Ipv4 NAT"
-                              checked={formik.values.ipv4_nat}
-                            />
-                          ),
-                        }),
-                      ]
-                    : []),
-
-                  ...(formik.values.bridge_mode !== "fan"
-                    ? [
-                        getConfigurationRow({
-                          formik: formik,
-                          name: "ipv6_address",
-                          label: "IPv6 Address",
-                          defaultValue: "auto",
-                          children: (
-                            <IpAddressSelector
-                              id="ipv6_address"
-                              address={formik.values.ipv6_address}
-                              setAddress={(value) => {
-                                formik.setFieldValue("ipv6_address", value);
-
-                                if (value === "none") {
-                                  const nullFields = [
-                                    "ipv6_nat",
-                                    "ipv6_dhcp",
-                                    "ipv6_dhcp_expiry",
-                                    "ipv6_dhcp_ranges",
-                                    "ipv6_dhcp_stateful",
-                                    "ipv6_ovn_ranges",
-                                  ];
-                                  nullFields.forEach((field) =>
-                                    formik.setFieldValue(field, undefined)
-                                  );
-                                }
-                              }}
-                            />
-                          ),
-                        }),
-                      ]
-                    : []),
-
-                  ...(formik.values.bridge_mode !== "fan" &&
-                  formik.values.ipv6_address !== "none"
-                    ? [
-                        getConfigurationRow({
-                          formik: formik,
-                          name: "ipv6_nat",
-                          label: "Ipv6 NAT",
-                          defaultValue: "",
-                          children: (
-                            <CheckboxInput
-                              label="Ipv6 NAT"
-                              checked={formik.values.ipv6_nat}
-                            />
-                          ),
-                        }),
-                      ]
-                    : []),
-
-                  ...(formik.values.bridge_mode === "fan"
-                    ? [
-                        getConfigurationRow({
-                          formik: formik,
-                          name: "fan_overlay_subnet",
-                          label: "Fan overlay subnet",
-                          help: "Subnet to use as the overlay for the FAN (CIDR)",
-                          defaultValue: "",
-                          children: <Input type="text" />,
-                        }),
-                      ]
-                    : []),
-
-                  ...(formik.values.bridge_mode === "fan"
-                    ? [
-                        getConfigurationRow({
-                          formik: formik,
-                          name: "fan_underlay_subnet",
-                          label: "Fan underlay subnet",
-                          help: "Subnet to use as the underlay for the FAN (use auto to use default gateway subnet) (CIDR)",
-                          defaultValue: "",
-                          children: <Input type="text" />,
-                        }),
-                      ]
-                    : []),
-                ]}
-              />
-            </React.Fragment>
+            <NetworkFormMain formik={formik} project={project} />
           )}
-          {section === BRIDGE && (
-            <ConfigurationTable
-              rows={[
-                getConfigurationRow({
-                  formik: formik,
-                  name: "bridge_mtu",
-                  label: "MTU",
-                  help: "Bridge MTU (default varies if tunnel or fan setup)",
-                  defaultValue: "",
-                  children: <Input type="text" />,
-                }),
-
-                getConfigurationRow({
-                  formik: formik,
-                  name: "bridge_hwaddr",
-                  label: "Hardware address",
-                  help: "MAC address for the bridge",
-                  defaultValue: "",
-                  children: <Input type="text" />,
-                }),
-
-                ...(formik.values.type === "bridge"
-                  ? [
-                      getConfigurationRow({
-                        formik: formik,
-                        name: "bridge_driver",
-                        label: "Bridge Driver",
-                        help: "Bridge driver: native or openvswitch",
-                        defaultValue: "",
-                        children: (
-                          <Select
-                            options={[
-                              {
-                                label: "Select option",
-                                value: "",
-                                disabled: true,
-                              },
-                              {
-                                label: "Native",
-                                value: "native",
-                              },
-                              {
-                                label: "Openvswitch",
-                                value: "openvswitch",
-                              },
-                            ]}
-                          />
-                        ),
-                      }),
-                    ]
-                  : []),
-              ]}
-            />
-          )}
-          {section === DNS && (
-            <ConfigurationTable
-              rows={[
-                getConfigurationRow({
-                  formik: formik,
-                  name: "dns_domain",
-                  label: "DNS domain",
-                  help: "Domain to advertise to DHCP clients and use for DNS resolution",
-                  defaultValue: "",
-                  children: <Input type="text" />,
-                }),
-
-                ...(formik.values.type === "bridge"
-                  ? [
-                      getConfigurationRow({
-                        formik: formik,
-                        name: "dns_mode",
-                        label: "DNS mode",
-                        defaultValue: "",
-                        children: (
-                          <Select
-                            options={[
-                              {
-                                label: "Select option",
-                                value: "",
-                                disabled: true,
-                              },
-                              {
-                                label: "None",
-                                value: "none",
-                              },
-                              {
-                                label: "Managed",
-                                value: "managed",
-                              },
-                              {
-                                label: "Dynamic",
-                                value: "dynamic",
-                              },
-                            ]}
-                          />
-                        ),
-                      }),
-                    ]
-                  : []),
-
-                getConfigurationRow({
-                  formik: formik,
-                  name: "dns_search",
-                  label: "DNS search",
-                  help: "Full comma-separated domain search list, defaulting to DNS domain value",
-                  defaultValue: "",
-                  children: <Textarea />,
-                }),
-              ]}
-            />
-          )}
-          {section === IPV4 && (
-            <ConfigurationTable
-              rows={[
-                getConfigurationRow({
-                  formik: formik,
-                  name: "ipv4_dhcp",
-                  label: "IPv4 DHCP",
-                  defaultValue: "",
-                  children: (
-                    <CheckboxInput
-                      label="IPv4 DHCP"
-                      checked={formik.values.ipv4_dhcp}
-                    />
-                  ),
-                }),
-
-                ...(formik.values.type !== "ovn" &&
-                (formik.values.ipv4_dhcp === true ||
-                  formik.values.ipv4_dhcp === undefined)
-                  ? [
-                      getConfigurationRow({
-                        formik: formik,
-                        name: "ipv4_dhcp_expiry",
-                        label: "IPv4 DHCP expiry",
-                        help: "When to expire DHCP leases",
-                        defaultValue: "",
-                        children: <Input type="text" />,
-                      }),
-
-                      getConfigurationRow({
-                        formik: formik,
-                        name: "ipv4_dhcp_ranges",
-                        label: "IPv4 DHCP ranges",
-                        help: "Comma-separated list of IP ranges to use for DHCP (FIRST-LAST format)",
-                        defaultValue: "",
-                        children: <Textarea />,
-                      }),
-                    ]
-                  : []),
-
-                ...(formik.values.type === "ovn"
-                  ? [
-                      getConfigurationRow({
-                        formik: formik,
-                        name: "ipv4_l3only",
-                        label: "IPv4 L3 only",
-                        defaultValue: "",
-                        children: (
-                          <CheckboxInput
-                            label="IPv4 L3 only"
-                            checked={formik.values.ipv4_l3only}
-                          />
-                        ),
-                      }),
-                    ]
-                  : []),
-
-                ...(formik.values.type !== "ovn" &&
-                formik.values.bridge_mode !== "fan"
-                  ? [
-                      getConfigurationRow({
-                        formik: formik,
-                        name: "ipv4_ovn_ranges",
-                        label: "IPv4 OVN ranges",
-                        help: "Comma-separated list of IPv4 ranges to use for child OVN network routers (FIRST-LAST format)",
-                        defaultValue: "",
-                        children: <Textarea />,
-                      }),
-                    ]
-                  : []),
-              ]}
-            />
-          )}
-          {section === IPV6 && (
-            <ConfigurationTable
-              rows={[
-                getConfigurationRow({
-                  formik: formik,
-                  name: "ipv6_dhcp",
-                  label: "IPv6 DHCP",
-                  defaultValue: "",
-                  children: (
-                    <CheckboxInput
-                      label="IPv6 DHCP"
-                      checked={formik.values.ipv6_dhcp}
-                    />
-                  ),
-                }),
-
-                ...(formik.values.ipv6_dhcp === true &&
-                formik.values.type !== "ovn"
-                  ? [
-                      getConfigurationRow({
-                        formik: formik,
-                        name: "ipv6_dhcp_expiry",
-                        label: "IPv6 DHCP expiry",
-                        help: "When to expire DHCP leases",
-                        defaultValue: "",
-                        children: <Input type="text" />,
-                      }),
-
-                      getConfigurationRow({
-                        formik: formik,
-                        name: "ipv6_dhcp_ranges",
-                        label: "IPv6 DHCP ranges",
-                        help: "Comma-separated list of IPv6 ranges to use for DHCP (FIRST-LAST format)",
-                        defaultValue: "",
-                        children: <Textarea />,
-                      }),
-                    ]
-                  : []),
-
-                ...(formik.values.ipv6_dhcp === true ||
-                formik.values.ipv4_dhcp === undefined
-                  ? [
-                      getConfigurationRow({
-                        formik: formik,
-                        name: "ipv6_dhcp_stateful",
-                        label: "IPv6 DHCP stateful",
-                        defaultValue: "",
-                        children: (
-                          <CheckboxInput
-                            label="IPv6 DHCP stateful"
-                            checked={formik.values.ipv6_dhcp_stateful}
-                          />
-                        ),
-                      }),
-                    ]
-                  : []),
-
-                ...(formik.values.type === "ovn"
-                  ? [
-                      getConfigurationRow({
-                        formik: formik,
-                        name: "ipv6_l3only",
-                        label: "IPv6 L3 only",
-                        defaultValue: "",
-                        children: (
-                          <CheckboxInput
-                            label="IPv6 L3 only"
-                            checked={formik.values.ipv6_l3only}
-                          />
-                        ),
-                      }),
-                    ]
-                  : []),
-
-                ...(formik.values.type !== "ovn"
-                  ? [
-                      getConfigurationRow({
-                        formik: formik,
-                        name: "ipv6_ovn_ranges",
-                        label: "IPv6 OVN ranges",
-                        help: "Comma-separated list of IPv6 ranges to use for child OVN network routers (FIRST-LAST format)",
-                        defaultValue: "",
-                        children: <Textarea />,
-                      }),
-                    ]
-                  : []),
-              ]}
-            />
-          )}
+          {section === BRIDGE && <NetworkFormBridge formik={formik} />}
+          {section === DNS && <NetworkFormDns formik={formik} />}
+          {section === IPV4 && <NetworkFormIpv4 formik={formik} />}
+          {section === IPV6 && <NetworkFormIpv6 formik={formik} />}
           {section === YAML_CONFIGURATION && (
             <YamlForm
               yaml={getYaml()}
